refactor : 주석 추가및 일부 로직 수정(#66)

See merge request s10-final/S10P31C204!239

+++ api/src/main/java/org/egglog/api/worktype/model/service/WorkTypeService.java
@@ -38,7 +38,6 @@ public class WorkTypeService {
         log.debug(" ==== ==== ==== [ 근무 타입 수정 서비스 실행 ] ==== ==== ====");
         WorkType workType = workTypeJpaRepository.findWithUserById(request.getWorkTypeId())
                 .orElseThrow(() -> new WorkTypeException(NO_EXIST_WORKTYPE));
-        if (workType.getWorkTag().name() != WorkTag.ETC.name()) throw new WorkTypeException(MODIFY_DEFAULT_TAG_DENIED);
         if (workType.getUser().getId() != loginUser.getId()) throw new WorkTypeException(ACCESS_DENIED);
         return workTypeJpaRepository.save(workType
                     .edit(request.getTitle(), request.getColor(), request.getWorkTypeImgUrl(), request.getStartTime(), request.getWorkTime()))
@@ -50,6 +49,7 @@ public class WorkTypeService {
         log.debug(" ==== ==== ==== [ 근무 타입 삭제 서비스 실행 ] ==== ==== ==== ");
         WorkType workType = workTypeJpaRepository.findWithUserById(workTypeId)
                 .orElseThrow(() -> new WorkTypeException(NO_EXIST_WORKTYPE));
+        if (workType.getWorkTag().equals(WorkTag.ETC.name())) throw new WorkTypeException(ACCESS_DENIED);
         if (workType.getUser().getId()!=loginUser.getId()) throw new WorkTypeException(ACCESS_DENIED);
         workTypeJpaRepository.delete(workType);
     }
